requirements.txt
fastapi==0.103.2
pydantic==2.4.2
pydantic_settings==2.0.3
uvicorn[standard]==0.23.2
aioboto3==11.3.0
google-cloud-firestore==2.12.0
motor==3.3.1
redis==4.6.0
fastapi-limiter==0.1.5
ujson==5.8.0
elastic-apm==6.19.0
strawberry-graphql[fastapi]==0.209.6
strawberry-graphql[cli]==0.209.6